2.1 changes
When the General node is in wake mode, all three nodes will be refreshed including the Climate and Security nodes
A new configuration parameter "Enable Security Commands" has been added
