+++ lib/tesla.js
@@ -234,12 +234,6 @@ module.exports = function(Polyglot, polyInterface) {
         '/api/1/vehicles/' + id + '/data_request/charge_state');
     }
 
-    // Get a vehicle GUI settings
-    async getVehicleGuiSettings(id) {
-      return await this.teslaApiGet(
-        '/api/1/vehicles/' + id + '/data_request/gui_settings');
-    }
-
     async wakeUp(id) {
       return await this.teslaApiPost('/api/1/vehicles/' + id + '/wake_up');
     }
