+++ profile/nls/en_US.txt
@@ -6,14 +6,23 @@ CMD-CONTROLLER-UPDATE_PROFILE-NAME = Update node server profile
 CMD-CONTROLLER-QUERY-NAME = Query node server
 
 ND-VEHICLE-NAME = Vehicle
-ND-VEHICLE-ICON = Electricity
+ND-VEHICLE-ICON = GenericRspCtl
 ST-V-ST-NAME = % State of charge
 ST-V-GV1-NAME = Battery range
+ST-V-GV2-NAME = Charge port door open
+ST-V-GV3-NAME = Charge port latch engaged
 ST-V-GV4-NAME = Charge enabled
 ST-V-GV5-NAME = Charging state
 ST-V-GV6-NAME = Fast charger present
 ST-V-GV7-NAME = Charge limit
+ST-V-GV8-NAME = Locked
+ST-V-GV9-NAME = % Sunroof
 ST-V-GV10-NAME = Odometer
+ST-V-GV11-NAME = Sentry mode on
+ST-V-GV12-NAME = Drivers side temp
+ST-V-GV13-NAME = Passengers side temp
+ST-V-GV14-NAME = Outside temp
+ST-V-GV17-NAME = Software Update Status
 ST-V-GV18-NAME = Vehicle Online
 ST-V-GV19-NAME = Timestamp
 ST-V-GV20-NAME = Car ID
@@ -21,6 +30,8 @@ ST-V-CC-NAME = Charger actual current
 ST-V-CV-NAME = Charger voltage
 ST-V-CPW-NAME = Charger power
 ST-V-ERR-NAME = Error
+ST-V-CLITEMP-NAME = Interior temperature
+ST-V-CLIEMD-NAME = Climate conditioning on
 
 
 CMD-V-DON-NAME = Charge On
@@ -29,81 +40,29 @@ CMD-V-WAKE-NAME = Wake
 CMD-V-LETSLEEP-NAME = Let vehicle sleep
 CMD-V-HORN-NAME = Horn
 CMD-V-FLASH-NAME = Flash lights
-CMD-V-CHARGE_SET_TO-NAME = Charge limit
-CMD-V-QUERY_NOW-NAME = Query
-
-EN_I_SOFTWARE_UPDATE_STATUS-0 = None Pending
-EN_I_SOFTWARE_UPDATE_STATUS-1 = Available 
-EN_I_SOFTWARE_UPDATE_STATUS-2 = Scheduled
-EN_I_SOFTWARE_UPDATE_STATUS-3 = Installing
-
-EN_I_OPEN_STATUS-0 = Closed
-EN_I_OPEN_STATUS-1 = Open
-
-EN_I_SENTRY_MODE-0 = Off
-EN_I_SENTRY_MODE-1 = On
-
-EN_I_CHARGE_PORT_DOOR-0 = Closed
-EN_I_CHARGE_PORT_DOOR-1 = Open
-
-EN_I_SEAT_HEATER_LEVEL-0 = Off
-EN_I_SEAT_HEATER_LEVEL-1 = Low 
-EN_I_SEAT_HEATER_LEVEL-2 = Medium
-EN_I_SEAT_HEATER_LEVEL-3 = Hot Buns
-
-ND-VEHSEC-NAME = Vehicle Security
-ND-VEHSEC-ICON = DoorLock
-ST-S-GV1-NAME = Charge port door
-ST-S-GV2-NAME = Charge port latch engaged
-ST-S-GV3-NAME = Frunk Status
-ST-S-GV4-NAME = Trunk Status
-ST-S-GV8-NAME = Locked
-ST-S-GV9-NAME = % Sunroof
-ST-S-GV11-NAME = Sentry mode on
-ST-S-GV17-NAME = Software Update Status
-ST-S-GV18-NAME = Vehicle Online
-ST-S-GV19-NAME = Timestamp
-ST-S-GV20-NAME = Car ID
-ST-S-ERR-NAME = Error
-
-CMD-S-QUERY_NOW-NAME = Query
-CMD-S-LOCK-NAME = Lock
-CMD-S-UNLOCK-NAME = Unlock
-CMD-S-SUNROOF_OPEN-NAME = Open roof
-CMD-S-SUNROOF_CLOSE-NAME = Close roof
-CMD-S-WINDOWS_VENT-NAME = Vent windows
-CMD-S-WINDOWS_CLOSE-NAME = Close windows
-CMD-S-TRUNK_OPEN-NAME = Open trunk
-CMD-S-FRUNK_OPEN-NAME = Open frunk
-CMD-S-CHARGE_PORT_DOOR-NAME = Charge port door
-CMD-S-SENTRY_MODE-NAME = Sentry mode
-CMD-S-START_SOFTWARE_UPDATE-NAME = Start software update
-
-
-ND-VEHICLECLIMATE-NAME = Vehicle Climate
-ND-VEHICLECLIMATE-ICON = Thermostat
-ST-C-GV12-NAME = Drivers side temp
-ST-C-GV13-NAME = Passengers side temp
-ST-C-GV14-NAME = Outside temp
-ST-C-GV15-NAME = Max Defrost
-ST-C-GV18-NAME = Vehicle Online
-ST-C-GV19-NAME = Timestamp
-ST-C-GV20-NAME = Car ID
-ST-C-ERR-NAME = Error
-ST-C-CLITEMP-NAME = Interior temperature
-ST-C-CLIEMD-NAME = Climate conditioning on
-
-
-CMD-C-QUERY_NOW-NAME = Query
-CMD-C-CLIMATE_ON-NAME = Conditioning On
-CMD-C-CLIMATE_OFF-NAME = Conditioning Off
-CMD-C-HEATED_SEAT_LEVEL_DRIVER-NAME = Driver's heated seat level
-CMD-C-HEATED_SEAT_LEVEL_PASSENGER-NAME = Passenger's heated seat level
-CMD-C-HEATED_SEAT_LEVEL_REAR_LEFT-NAME = Rear left heated seat level
-CMD-C-HEATED_SEAT_LEVEL_REAR_CENTER-NAME = Rear center heated seat level
-CMD-C-HEATED_SEAT_LEVEL_REAR_RIGHT-NAME = Rear right heated seat level
-CMD-C-MAX_DEFROST_ON-NAME = Max defrost on
-CMD-C-MAX_DEFROST_OFF-NAME = Max defrost off
-CMD-C-CLIMATE_TEMP_SETTING_DRIVER-NAME = Driver's climate control temp
-CMD-C-CLIMATE_TEMP_SETTING_PASSENGER-NAME = Passenger's climate control temp
-
+CMD-V-LOCK-NAME = Lock
+CMD-V-UNLOCK-NAME = Unlock
+CMD-V-SUNROOF_OPEN-NAME = Open roof
+CMD-V-SUNROOF_CLOSE-NAME = Close roof
+CMD-V-PORT_OPEN-NAME = Open port
+CMD-V-PORT_CLOSE-NAME = Close port
+CMD-V-CHARGE_SET_TO-NAME = Set charge limit
+CMD-V-QUERY-NAME = Query
+CMD-V-CLIMATE_ON-NAME = Conditioning On
+CMD-V-CLIMATE_OFF-NAME = Conditioning Off
+CMD-V-WINDOWS_VENT-NAME = Vent windows
+CMD-V-WINDOWS_CLOSE-NAME = Close windows
+CMD-V-TRUNK_OPEN-NAME = Open trunk
+CMD-V-FRUNK_OPEN-NAME = Open frunk
+CMD-V-HEATED_SEAT_LEVEL_DRIVER-NAME = Driver's heated seat level
+CMD-V-HEATED_SEAT_LEVEL_PASSENGER-NAME = Passenger's heated seat level
+CMD-V-HEATED_SEAT_LEVEL_REAR_LEFT-NAME = Rear left heated seat level
+CMD-V-HEATED_SEAT_LEVEL_REAR_CENTER-NAME = Rear center heated seat level
+CMD-V-HEATED_SEAT_LEVEL_REAR_RIGHT-NAME = Rear right heated seat level
+CMD-V-SENTRY_MODE_ON-NAME = Sentry mode on
+CMD-V-SENTRY_MODE_OFF-NAME = Sentry mode off
+CMD-V-MAX_DEFROST_ON-NAME = Max defrost on
+CMD-V-MAX_DEFROST_OFF-NAME = Max defrost off
+CMD-V-START_SOFTWARE_UPDATE-NAME = Start software update
+CMD-V-CLIMATE_TEMP_SETTING_DRIVER-NAME = Driver's climate control temperature
+CMD-V-CLIMATE_TEMP_SETTING_PASSENGER-NAME = Passenger's climate control temperature
